requirements.txt
Flask==1.1.1
Jinja2==2.11.1
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.13
Flask-WTF==0.14.3
WTForms==2.2.1
PyMySQL==0.9.3
flask-bcrypt-0.7.1Flask-Login==0.5.0
flask-login==0.5.0
gunicorn==0.0.0
pytest==5.4.0
pytest-cov==2.8.1

